+++ src/main/java/org/example/final_project/repositories/PostRepository.java
@@ -7,15 +7,15 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface PostRepository extends JpaRepository<Post, Integer> {
+public interface PostRepository extends JpaRepository<Post, Long> {
 	
 	@Query("select p from Post p where p.user.id=?1")
-	public List<Post> findByUserId (Integer userId);
+	public List<Post> findByUserId (Long userId);
 	
     @Query("SELECT p FROM Post p WHERE p.user.id IN :users ORDER BY p.createdAt DESC")
-    public List<Post> findAllPostByUserIds(@Param("users") List<Integer> userIds);
+    public List<Post> findAllPostByUserIds(@Param("users") List<Long> userIds);
 
     @Query("SELECT p FROM Post p WHERE p.user.id IN :users ORDER BY p.createdAt DESC")
-    public List<Post> findAllPostByUserIdsSortedByDateDesc(@Param("users") List<Integer> userIds);
+    public List<Post> findAllPostByUserIdsSortedByDateDesc(@Param("users") List<Long> userIds);
 
 }
