+++ src/main/java/org/example/final_project/services/PostService.java
@@ -10,26 +10,26 @@ import java.util.List;
 
 public interface PostService {
 
-	public Post createPost(Post post, Integer userId) throws UserException;
+	public Post createPost(Post post, Long userId) throws UserException;
 	
-	public Post editPost(Post post,Integer userId) throws PostException;
+	public Post editPost(Post post,Long userId) throws PostException;
 	
-	public String deletePost(Integer postId, Integer userId) throws UserException,PostException;
+	public String deletePost(Long postId, Long userId) throws UserException,PostException;
 	
-	public List<Post> findPostByUserId(Integer userId) throws UserException;
+	public List<Post> findPostByUserId(Long userId) throws UserException;
 	
-	public Post findePostById(Integer postId) throws PostException;
+	public Post findePostById(Long postId) throws PostException;
 	
 	public List<Post> findAllPost() throws PostException;
 	
-	public List<Post> findAllPostByUserIds(List<Integer> userIds) throws PostException, UserException;
+	public List<Post> findAllPostByUserIds(List<Long> userIds) throws PostException, UserException;
 	
-	public String savedPost(Integer postId,Integer userId) throws PostException, UserException;
+	public String savedPost(Long postId,Long userId) throws PostException, UserException;
 	
-	public String unSavePost(Integer postId,Integer userId) throws PostException, UserException;
+	public String unSavePost(Long postId,Long userId) throws PostException, UserException;
 		
-	public Post likePost(Integer postId ,Integer userId) throws UserException, PostException;
+	public Post likePost(Long postId ,Long userId) throws UserException, PostException;
 	
-	public Post unLikePost(Integer postId ,Integer userId) throws UserException, PostException;
+	public Post unLikePost(Long postId ,Long userId) throws UserException, PostException;
 	
 }
